requirements.txt
numpy==1.19.3
opencv-python==4.4.0.46
python-rtmidi==1.4.6
